Fix some more GCC warnings
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7045 3db68df2-c116-0410-a063-a993310a9797

+++ source/network/NetClient.cpp
@@ -250,7 +250,7 @@ bool CNetClient::OnError( void* pContext, CFsmEvent* pEvent )
 	if ( !pEvent || !pContext ) return false;
 
 	// Error event?
-	if ( pEvent->GetType() != NMT_ERROR ) return true;
+	if ( pEvent->GetType() != (uint)NMT_ERROR ) return true;
 
 	CNetClient*	pClient = ( CNetClient* )( ( FsmActionCtx* )pContext )->pHost;
 	assert( pClient );
@@ -364,7 +364,7 @@ bool CNetClient::OnAuthenticate( void* pContext, CFsmEvent* pEvent )
 	assert( pClient );
 	assert( pSession );
 
-	if ( pEvent->GetType() ==  NMT_ERROR )
+	if ( pEvent->GetType() == (uint)NMT_ERROR )
 	{
 		// return CNetClient::OnError( pContext, pEvent );
 	}
